+++ clash-lib/src/Clash/Backend/Verilog.hs
@@ -69,6 +69,7 @@ import           Clash.Netlist.BlackBox.Util
 import           Clash.Netlist.Id                     (IdType (..), mkBasicId')
 import           Clash.Netlist.Types                  hiding (_intWidth, intWidth)
 import           Clash.Netlist.Util                   hiding (mkIdentifier, extendIdentifier)
+import           Clash.Signal.Internal                (ActiveEdge (..))
 import           Clash.Util
   (SrcSpan, noSrcSpan, curLoc, traceIf, (<:>), on, first, indexNote)
 
@@ -314,9 +315,10 @@ uselibs xs = line <>
   indent 2 (string "`uselib" <+> (hsep (mapM (\l -> ("lib=" <> string l)) xs)))
   <> line <> line
 
-wireOrRegDoc :: WireOrReg -> VerilogM Doc
-wireOrRegDoc Wire = "wire"
-wireOrRegDoc Reg  = "reg"
+wireRegFileDoc :: WireOrReg -> (Either Identifier HWType) -> VerilogM Doc
+wireRegFileDoc _    (Right FileType) = "integer"
+wireRegFileDoc Wire _                = "wire"
+wireRegFileDoc Reg  _                = "reg"
 
 addSeen :: Component -> VerilogM ()
 addSeen c = do
@@ -332,6 +334,7 @@ verilogType t = case t of
   Reset {} -> emptyDoc
   Bit      -> emptyDoc
   Bool     -> emptyDoc
+  FileType -> emptyDoc
   _        -> brackets (int (typeSize t -1) <> colon <> int 0)
 
 sigDecl :: VerilogM Doc -> HWType -> VerilogM Doc
@@ -387,7 +390,7 @@ renderAttr (Attr'        key      ) = pack $ key
 
 decl :: Declaration -> VerilogM (Maybe Doc)
 decl (NetDecl' noteM wr id_ tyE iEM) =
-  Just A.<$> maybe id addNote noteM (addAttrs attrs (wireOrRegDoc wr <+> tyDec tyE))
+  Just A.<$> maybe id addNote noteM (addAttrs attrs (wireRegFileDoc wr tyE <+> tyDec tyE))
   where
     tyDec (Left  ty) = stringS ty <+> stringS id_ <> iE
     tyDec (Right ty) = sigDecl (stringS id_) ty <> iE
@@ -542,7 +545,65 @@ inst_ (InstDecl _ _ nm lbl ps pms) = fmap Just $
 inst_ (BlackBoxD _ libs imps inc bs bbCtx) =
   fmap Just (Mon (column (renderBlackBox libs imps inc bs bbCtx)))
 
+inst_ (Seq ds) = Just <$> seqs ds
+
 inst_ (NetDecl' {}) = return Nothing
+
+seq_ :: Seq -> VerilogM Doc
+seq_ (AlwaysClocked edge clk ds) =
+  "always @" <>
+    parens (case edge of {Rising -> "posedge"; _ -> "negedge"} <+>
+            expr_ False clk) <+> "begin" <> line <>
+    indent 2 (seqs ds) <> line <>
+  "end"
+
+seq_ (Initial ds) =
+  "initial begin" <> line <>
+  indent 2 (seqs ds) <> line <>
+  "end"
+
+seq_ (AlwaysComb ds) =
+  "always @* begin" <> line <>
+  indent 2 (seqs ds) <> line <>
+  "end"
+
+seq_ (Branch scrut scrutTy es) =
+    "case" <> parens (expr_ True scrut) <> line <>
+      (indent 2 $ vcat $ conds es) <> line <>
+    "endcase"
+   where
+    conds :: [(Maybe Literal,[Seq])] -> VerilogM [Doc]
+    conds [] =
+      return []
+    conds [(_,sq)] =
+      ("default" <+> colon <+> "begin" <> line <>
+        indent 2 (seqs sq) <> line <>
+      "end") <:> return []
+    conds ((Nothing,sq):_) =
+      ("default" <+> colon <+> "begin" <> line <>
+        indent 2 (seqs sq) <> line <>
+      "end") <:> return []
+    conds ((Just c ,sq):es') =
+      (exprLitV (Just (scrutTy,conSize scrutTy)) c <+> colon <+> "begin" <> line <>
+        indent 2 (seqs sq) <> line <>
+      "end") <:> conds es'
+
+seq_ (SeqDecl sd) = case sd of
+  Assignment id_ e ->
+    stringS id_ <+> equals <+> expr_ False e <> semi
+
+  BlackBoxD {} ->
+    fromMaybe <$> emptyDoc <*> inst_ sd
+
+  Seq ds ->
+    seqs ds
+
+  _ -> error ("seq_: " ++ show sd)
+
+seqs :: [Seq] -> VerilogM Doc
+seqs [] = emptyDoc
+seqs (SeqDecl (TickDecl id_):ds) = "//" <+> stringS id_ <> line <> seqs ds
+seqs (d:ds) = seq_ d <> line <> line <> seqs ds
 
 -- | Calculate the beginning and end index into a variable, to get the
 -- desired field.