+++ clash-lib/src/Clash/Normalize/Transformations.hs
@@ -109,7 +109,7 @@ import           Clash.Core.Type             (Type, TypeView (..), applyFunTy,
                                               isPolyFunCoreTy, isClassTy,
                                               normalizeType, splitFunForallTy,
                                               splitFunTy,
-                                              tyView, mkPolyFunTy)
+                                              tyView, mkPolyFunTy, coreView)
 import           Clash.Core.TyCon            (TyConMap, tyConDataCons)
 import           Clash.Core.Util
   (isCon, isFun, isLet, isPolyFun, isPrim,
@@ -691,6 +691,10 @@ caseOneAlt e@(Case _ _ [(pat,altE)]) = case pat of
     | otherwise
     -> return e
 
+caseOneAlt (Case _ _ alts@((_,alt):_:_))
+  | all ((== alt) . snd) (tail alts)
+  = changed alt
+
 caseOneAlt e = return e
 
 -- | Bring an application of a DataCon or Primitive in ANF, when the argument is
@@ -776,6 +780,7 @@ removeUnusedExpr _ e@(collectArgsTicks -> (p@(Prim nm pInfo),args,ticks)) = do
       let usedArgs | isFromInt pNm
                    = [0,1,2]
                    | nm `elem` ["Clash.Annotations.BitRepresentation.Deriving.dontApplyInHDL"
+                               ,"Clash.Sized.Vector.splitAt"
                                ]
                    = [0,1]
                    | otherwise
@@ -1390,6 +1395,36 @@ type NormRewriteW = Transform (StateT ([LetBinding],InScopeSet) (RewriteMonad No
 tellBinders :: Monad m => [LetBinding] -> StateT ([LetBinding],InScopeSet) m ()
 tellBinders bs = modify ((bs ++) *** (`extendInScopeSetList` (map fst bs)))
 
+-- | See Note [ANF InScopeSet]; only extends the inscopeset
+notifyBinders :: Monad m => [LetBinding] -> StateT ([LetBinding],InScopeSet) m ()
+notifyBinders bs = modify (second (`extendInScopeSetList` (map fst bs)))
+
+-- | Is the given type IO-like
+isSimIOTy
+  :: TyConMap
+  -> Type
+  -- ^ Type to check for IO-likeness
+  -> Bool
+isSimIOTy tcm ty = case tyView (coreView tcm ty) of
+  TyConApp tcNm args
+    | nameOcc tcNm == "Clash.Explicit.SimIO.SimIO"
+    -> True
+    | nameOcc tcNm == "GHC.Prim.(#,#)"
+    , [_,_,st,_] <- args
+    -> isStateTokenTy tcm st
+  FunTy _ res -> isSimIOTy tcm res
+  _ -> False
+
+-- | Is the given type the state token
+isStateTokenTy
+  :: TyConMap
+  -> Type
+  -- ^ Type to check for state tokenness
+  -> Bool
+isStateTokenTy tcm ty = case tyView (coreView tcm ty) of
+  TyConApp tcNm _ -> nameOcc tcNm == "GHC.Prim.State#"
+  _ -> False
+
 -- | Turn an expression into a modified ANF-form. As opposed to standard ANF,
 -- constants do not become let-bound.
 makeANF :: HasCallStack => NormRewrite
@@ -1414,9 +1449,10 @@ makeANF ctx@(TransformContext is0 _) e0
     --
     -- See also Note [ANF InScopeSet]
     let (is2,e1) = freshenTm is0 e0
-    (e2,(bndrs,_)) <- runStateT (bottomupR collectANF ctx e1) ([],is2)
+    ((e2,(bndrs,_)),Monoid.getAny -> hasChanged) <-
+      listen (runStateT (bottomupR collectANF ctx e1) ([],is2))
     case bndrs of
-      [] -> return e0
+      [] -> if hasChanged then return e2 else return e0
       _  -> do
         let (e3,ticks) = collectTicks e2
             (srcTicks,nmTicks) = partitionTicks ticks
@@ -1458,6 +1494,19 @@ makeANF ctx@(TransformContext is0 _) e0
 -- So we start out with an InScopeSet that satisfies points 1 and 2, now every
 -- time we create a new binder we must add it to the InScopeSet to satisfy
 -- point 3.
+--
+-- Note [ANF no let-bind]
+--
+-- | Do not let-bind:
+--
+-- 1. Arguments with an untranslatable type: untranslatable expressions
+--    should be propagated down as far as possible
+--
+-- 2. Local variables or constants: they don't add any work, so no reason
+--    to let-bind to enable sharing
+--
+-- 3. IO actions, the translation of IO actions to sequential HDL constructs
+--    depends on IO actions to be propagated down as far as possible.
 collectANF :: HasCallStack => NormRewriteW
 collectANF ctx e@(App appf arg)
   | (conVarPrim, _) <- collectArgs e
@@ -1466,8 +1515,9 @@ collectANF ctx e@(App appf arg)
     untranslatable <- lift (isUntranslatable False arg)
     let localVar   = isLocalVar arg
     constantNoCR   <- lift (isConstantNotClockReset arg)
-    case (untranslatable,localVar || constantNoCR,arg) of
-      (False,False,_) -> do
+    -- See Note [ANF no let-bind]
+    case (untranslatable,localVar || constantNoCR, isSimBind conVarPrim,arg) of
+      (False,False,False,_) -> do
         tcm <- Lens.view tcCache
         -- See Note [ANF InScopeSet]
         is1   <- Lens.use _2
@@ -1475,24 +1525,31 @@ collectANF ctx e@(App appf arg)
         -- See Note [ANF InScopeSet]
         tellBinders [(argId,arg)]
         return (App appf (Var argId))
-      (True,False,Letrec binds body) -> do
+      (True,False,_,Letrec binds body) -> do
         tellBinders binds
         return (App appf body)
       _ -> return e
+ where
+  isSimBind (Prim nm _) = nm == "Clash.Explicit.SimIO.bindSimIO#"
+  isSimBind _ = False
 
 collectANF _ (Letrec binds body) = do
-  tellBinders binds
+  tcm <- Lens.view tcCache
+  let isSimIO = isSimIOTy tcm (termType tcm body)
   untranslatable <- lift (isUntranslatable False body)
   let localVar = isLocalVar body
-  if localVar || untranslatable
-    then return body
+  -- See Note [ANF no let-bind]
+  if localVar || untranslatable || isSimIO
+    then do
+      tellBinders binds
+      return body
     else do
-      tcm <- Lens.view tcCache
       -- See Note [ANF InScopeSet]
       is1 <- Lens.use _2
       argId <- lift (mkTmBinderFor is1 tcm (mkUnsafeSystemName "result" 0) body)
       -- See Note [ANF InScopeSet]
       tellBinders [(argId,body)]
+      tellBinders binds
       return (Var argId)
 
 -- TODO: The code below special-cases ANF for the ':-' constructor for the
@@ -1518,36 +1575,41 @@ collectANF ctx (Case subj ty alts) = do
     let localVar = isLocalVar subj
     let isConstantSubj = isConstant subj
 
-    subj' <- if localVar || isConstantSubj
-      then return subj
+    (subj',subjBinders) <- if localVar || isConstantSubj
+      then return (subj,[])
       else do
         tcm <- Lens.view tcCache
         -- See Note [ANF InScopeSet]
         is1 <- Lens.use _2
         argId <- lift (mkTmBinderFor is1 tcm (mkDerivedName ctx "case_scrut") subj)
         -- See Note [ANF InScopeSet]
-        tellBinders [(argId,subj)]
-        return (Var argId)
+        notifyBinders [(argId,subj)]
+        return (Var argId,[(argId,subj)])
 
-    alts' <- mapM (doAlt subj') alts
+    tcm <- Lens.view tcCache
+    let isSimIOAlt = isSimIOTy tcm ty
+
+    alts' <- mapM (doAlt isSimIOAlt subj') alts
+    tellBinders subjBinders
 
     case alts' of
       [(DataPat _ [] xs,altExpr)]
-        | xs `localIdsDoNotOccurIn` altExpr
+        | xs `localIdsDoNotOccurIn` altExpr || isSimIOAlt
         -> return altExpr
       _ -> return (Case subj' ty alts')
   where
     doAlt
-      :: Term -> (Pat,Term)
+      :: Bool -> Term -> (Pat,Term)
       -> StateT ([LetBinding],InScopeSet) (RewriteMonad NormalizeState)
                 (Pat,Term)
-    doAlt subj' alt@(DataPat dc exts xs,altExpr) | not (bindsExistentials exts xs) = do
+    doAlt isSimIOAlt subj' alt@(DataPat dc exts xs,altExpr) | not (bindsExistentials exts xs) = do
       let lv = isLocalVar altExpr
       patSels <- Monad.zipWithM (doPatBndr subj' dc) xs [0..]
       let altExprIsConstant = isConstant altExpr
       let usesXs (Var n) = any (== n) xs
           usesXs _       = False
-      if (lv && (not (usesXs altExpr) || length alts == 1)) || altExprIsConstant
+      -- See [ANF no let-bind]
+      if or [isSimIOAlt, lv && (not (usesXs altExpr) || length alts == 1), altExprIsConstant]
         then do
           -- See Note [ANF InScopeSet]
           tellBinders patSels
@@ -1558,13 +1620,14 @@ collectANF ctx (Case subj ty alts) = do
           is1 <- Lens.use _2
           altId <- lift (mkTmBinderFor is1 tcm (mkDerivedName ctx "case_alt") altExpr)
           -- See Note [ANF InScopeSet]
-          tellBinders ((altId,altExpr):patSels)
+          tellBinders (patSels ++ [(altId,altExpr)])
           return (DataPat dc exts xs,Var altId)
-    doAlt _ alt@(DataPat {}, _) = return alt
-    doAlt _ alt@(pat,altExpr) = do
+    doAlt _ _ alt@(DataPat {}, _) = return alt
+    doAlt isSimIOAlt _ alt@(pat,altExpr) = do
       let lv = isLocalVar altExpr
       let altExprIsConstant = isConstant altExpr
-      if lv || altExprIsConstant
+      -- See [ANF no let-bind]
+      if isSimIOAlt || lv || altExprIsConstant
         then return alt
         else do
           tcm <- Lens.view tcCache
@@ -2218,6 +2281,7 @@ disjointExpressionConsolidation _ e = return e
 --     a HDL declaration)
 --   * a projection case-expression (1 alternative)
 --   * a data constructor
+--   * I/O actions
 inlineCleanup :: HasCallStack => NormRewrite
 inlineCleanup (TransformContext is0 _) (Letrec binds body) = do
   prims <- Lens.use (extra.primitives)
@@ -2255,11 +2319,26 @@ inlineCleanup (TransformContext is0 _) (Letrec binds body) = do
             , Just occ <- lookupVarEnv id_ allOccs
             , occ < 2
             -> True
+            | otherwise
+            -> nm `elem` ["Clash.Explicit.SimIO.bindSimIO#"]
           Case _ _ [_] -> True
           Data _ -> True
+          Case _ aTy (_:_:_)
+            | TyConApp (nameOcc -> "Clash.Explicit.SimIO.SimIO") _ <- tyView aTy
+            -> True
           _ -> False
       | id_ `notElemVarSet` bodyFVs
       = case tm of
+          Prim nm _
+            | nm `elem` [ "Clash.Explicit.SimIO.openFile"
+                        , "Clash.Explicit.SimIO.fgetc"
+                        , "Clash.Explicit.SimIO.feof"
+                        ]
+            , Just occ <- lookupVarEnv id_ allOccs
+            , occ < 2
+            -> True
+            | otherwise
+            -> nm `elem` ["Clash.Explicit.SimIO.bindSimIO#"]
           Case _ _ [(DataPat dcE _ _,_)]
             -> let nm = (nameOcc (dcName dcE))
                in -- Inlines WW projection that exposes internals of the BitVector types
@@ -2267,6 +2346,9 @@ inlineCleanup (TransformContext is0 _) (Letrec binds body) = do
                   nm == "Clash.Sized.Internal.BitVector.Bit" ||
                   -- Inlines projections out of constraint-tuples (e.g. HiddenClockReset)
                   "GHC.Classes" `Text.isPrefixOf` nm
+          Case _ aTy (_:_:_)
+            | TyConApp (nameOcc -> "Clash.Explicit.SimIO.SimIO") _ <- tyView aTy
+            -> True
           _ -> False
 
     isInteresting _ _ _ _ = False