+++ testsuite/Main.hs
@@ -73,6 +73,19 @@ runClashTest = defaultMain $ clashTestRoot
         , topEntity=TopEntity "test_i2c"
         , hdlSim=False
         }
+      , runTest "I2Ctest" def {
+          entities=Entities [ ".." </> "I2C" </> "i2c"
+                            , ".." </> "I2C" </> "bitmaster"
+                            , ".." </> "I2C" </> "bytemaster"
+                            , "configi2c"
+                            , "slave"
+                            , "system"
+                            ]
+        , topEntity=TopEntity "system"
+        , hdlTargets=[Verilog]
+        , hdlSim=True
+        , stdErrEmptyFail=False
+        }
       ]
     ]
   , clashTestGroup "tests"
@@ -327,6 +340,9 @@ runClashTest = defaultMain $ clashTestRoot
           ]
 
         , runFailingTest ("tests" </> "shouldfail" </> "Signal") allTargets [] "MAC" (Just "Couldn't instantiate blackbox for Clash.Signal.Internal.register#")
+        ]
+      , clashTestGroup "SimIO"
+        [ runTest "Test00" def {hdlTargets=[Verilog], stdErrEmptyFail=False, topEntity=TopEntity "topEntity"}
         ]
       , clashTestGroup "SynthesisAttributes"
         [ outputTest ("tests" </> "shouldwork" </> "SynthesisAttributes") allTargets [] [] "Simple"  "main"